Change StockId type from int to long

+++ Models/StockHistory.cs
@@ -1,7 +1,7 @@
 public class StockHistory
 {
     public int Id { get; set; }
-    public int StockId { get; set; }
+    public long StockId { get; set; }
     public DateTime Timestamp { get; set; }
     public decimal Price { get; set; }
     public decimal Quantity { get; set; }
